backend/requirements.txt
flask
octokitpy
sqlalchemy
psycopg2
Flask-SQLAlchemy
Flask-Migrate
SQLAlchemy-Searchable
SQLAlchemy-Utils